+++ test/Data/OpenApi/CommonTestTypes.hs
@@ -277,6 +277,7 @@ characterSchemaJSON = [aesonQQ|
         "tag",
         "contents"
       ],
+      "title": "PC",
       "type": "object",
       "properties": {
         "tag": {
@@ -296,6 +297,7 @@ characterSchemaJSON = [aesonQQ|
         "npcPosition",
         "tag"
       ],
+      "title": "NPC",
       "type": "object",
       "properties": {
         "tag": {
@@ -312,8 +314,7 @@ characterSchemaJSON = [aesonQQ|
         }
       }
     }
-  ],
-  "type": "object"
+  ]
 }
 
 |]
@@ -327,6 +328,7 @@ characterInlinedSchemaJSON = [aesonQQ|
         "tag",
         "contents"
       ],
+      "title": "PC",
       "type": "object",
       "properties": {
         "tag": {
@@ -368,6 +370,7 @@ characterInlinedSchemaJSON = [aesonQQ|
         "npcPosition",
         "tag"
       ],
+      "title": "NPC",
       "type": "object",
       "properties": {
         "tag": {
@@ -398,8 +401,7 @@ characterInlinedSchemaJSON = [aesonQQ|
         }
       }
     }
-  ],
-  "type": "object"
+  ]
 }
 |]
 
@@ -412,6 +414,7 @@ characterInlinedPlayerSchemaJSON = [aesonQQ|
         "tag",
         "contents"
       ],
+      "title": "PC",
       "type": "object",
       "properties": {
         "tag": {
@@ -439,6 +442,7 @@ characterInlinedPlayerSchemaJSON = [aesonQQ|
         "npcPosition",
         "tag"
       ],
+      "title": "NPC",
       "type": "object",
       "properties": {
         "tag": {
@@ -455,8 +459,7 @@ characterInlinedPlayerSchemaJSON = [aesonQQ|
         }
       }
     }
-  ],
-  "type": "object"
+  ]
 }
 |]
 
@@ -639,6 +642,7 @@ lightSchemaJSON = [aesonQQ|
       "required": [
         "tag"
       ],
+      "title": "NoLight",
       "type": "object",
       "properties": {
         "tag": {
@@ -654,6 +658,7 @@ lightSchemaJSON = [aesonQQ|
         "tag",
         "contents"
       ],
+      "title": "LightFreq",
       "type": "object",
       "properties": {
         "tag": {
@@ -673,6 +678,7 @@ lightSchemaJSON = [aesonQQ|
         "tag",
         "contents"
       ],
+      "title": "LightColor",
       "type": "object",
       "properties": {
         "tag": {
@@ -691,6 +697,7 @@ lightSchemaJSON = [aesonQQ|
         "waveLength",
         "tag"
       ],
+      "title": "LightWaveLength",
       "type": "object",
       "properties": {
         "tag": {
@@ -705,8 +712,7 @@ lightSchemaJSON = [aesonQQ|
         }
       }
     }
-  ],
-  "type": "object"
+  ]
 }
 |]
 
@@ -718,6 +724,7 @@ lightInlinedSchemaJSON = [aesonQQ|
       "required": [
         "tag"
       ],
+      "title": "NoLight",
       "type": "object",
       "properties": {
         "tag": {
@@ -733,6 +740,7 @@ lightInlinedSchemaJSON = [aesonQQ|
         "tag",
         "contents"
       ],
+      "title": "LightFreq",
       "type": "object",
       "properties": {
         "tag": {
@@ -752,6 +760,7 @@ lightInlinedSchemaJSON = [aesonQQ|
         "tag",
         "contents"
       ],
+      "title": "LightColor",
       "type": "object",
       "properties": {
         "tag": {
@@ -775,6 +784,7 @@ lightInlinedSchemaJSON = [aesonQQ|
         "waveLength",
         "tag"
       ],
+      "title": "LightWaveLength",
       "type": "object",
       "properties": {
         "tag": {
@@ -789,8 +799,7 @@ lightInlinedSchemaJSON = [aesonQQ|
         }
       }
     }
-  ],
-  "type": "object"
+  ]
 }
 |]
 
@@ -919,6 +928,7 @@ predicateSchemaDeclareJSON = [aesonQQ|
             "tag": { "enum": ["PredicateNoun"], "type": "string" }
           },
           "required": ["tag", "contents"],
+          "title": "PredicateNoun",
           "type": "object"
         },
         {
@@ -927,10 +937,10 @@ predicateSchemaDeclareJSON = [aesonQQ|
             "tag": { "enum": ["PredicateOmitted"], "type": "string" }
           },
           "required": ["tag", "contents"],
+          "title": "PredicateOmitted",
           "type": "object"
         }
-      ],
-      "type": "object"
+      ]
     },
     "Noun": {
       "properties": {
@@ -959,6 +969,7 @@ predicateSchemaDeclareJSON = [aesonQQ|
             "tag": { "enum": ["ModifierNoun"], "type": "string" }
           },
           "required": ["tag", "contents"],
+          "title": "ModifierNoun",
           "type": "object"
         },
         {
@@ -967,10 +978,10 @@ predicateSchemaDeclareJSON = [aesonQQ|
             "tag": { "enum": ["ModifierOmitted"], "type": "string" }
           },
           "required": ["tag", "contents"],
+          "title": "ModifierOmitted",
           "type": "object"
         }
-      ],
-      "type": "object"
+      ]
     },
     "Omitted": {
       "properties": {